conversational_api/requirements.txt
beautifulsoup4==4.12.3
dj_database_url==2.2.0
Django==5.1.1
djangorestframework==3.15.2
drf_nested_routers==0.94.1
drf_yasg==1.21.7
python-dotenv==1.0.1
Requests==2.32.3
